refactor elm creation into an ElmFactory class
